# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ccShield backend",
    "description": "Bilibili live-room moderator tool — backend API.",
    "version": "2.0.4"
  },
  "paths": {
...
    "/api/auth/qr/poll": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Poll B站 for the QR-login state",
        "description": "Poll B站 for the current state of an in-flight QR login.\n\nTranslates the B站 error codes into a single ``status`` field:\n\n- :class:`QrAwaitingScanError`    → ``scanning``\n- :class:`QrAwaitingConfirmError` → ``confirmed``\n- :class:`QrExpiredError`         → ``expired``\n- success path                    → ``success`` + persist + state refresh\n\nAny other exception is re-raised — those signal a real backend\nproblem (network down, B站 format change) and the client should\nsurface them as 5xx.",
        "operationId": "qr_poll_route_api_auth_qr_poll_get",
        "parameters": [
          {
            "name": "qrcode_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Qrcode Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QrPollResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/manual": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Save user-pasted cookies (Plan B fallback)",
        "description": "Validate pasted SESSDATA / bili_jct / buvid3 and persist on success.\n\nOn :class:`LoginIncompleteError` (the pasted cookies fail the\n``/nav`` validation or the response is malformed) we surface\nHTTP 400. The caller's existing ``.env`` is left untouched — the\n:func:`save_cookies_manual` helper does the same guarantee at the\npersistence layer.\n\nIf the user omitted ``buvid3`` in the request body,\nfetch one from ``/x/frontend/finger/spi`` and thread it through to\nboth ``save_cookies_manual`` and ``mark_authenticated_after_login``.\nIf the user did provide a buvid3 we trust it verbatim (it's the\ncookie they extracted from a real browser session, which is what\nB站 actually expects). fetch_buvid3 is best-effort and never raises.",
        "operationId": "manual_cookies_api_auth_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualCookiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualCookiesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ManualCookiesRequest": {
        "properties": {
          "sessdata": {
            "type": "string",
            "minLength": 1,
            "title": "Sessdata"
          },
          "bili_jct": {
            "type": "string",
            "minLength": 1,
            "title": "Bili Jct"
          },
          "buvid3": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buvid3"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessdata",
          "bili_jct"
        ],
        "title": "ManualCookiesRequest",
        "description": "Request body for ``POST /api/auth/manual`` (Plan B)."
      },
      "ManualCookiesResponse": {
        "properties": {
          "uname": {
            "type": "string",
            "title": "Uname"
          },
          "mid": {
            "type": "integer",
            "title": "Mid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "uname",
          "mid"
        ],
        "title": "ManualCookiesResponse",
        "description": "Response body for ``POST /api/auth/manual`` on success."
      },
...
      "QrPollResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "scanning",
              "confirmed",
              "expired",
              "success"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "QrPollResponse",
        "description": "Response body for ``GET /api/auth/qr/poll``.\n\n``status`` is one of:\n\n- ``\"scanning\"`` — B站 code 86101; the user has not scanned the QR yet.\n- ``\"confirmed\"`` — B站 code 86090; scanned, awaiting phone confirm.\n- ``\"expired\"``  — B站 code 86038; the QR expired, regenerate.\n- ``\"success\"``  — login complete; cookies are now persisted."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}